Add test script to build and test the auth server
The test.sh script builds the auth server and runs it with a
testconfig.yml configuration file that has some users and ACLs defined.
Then it fires requests against the started server to test if the
behavior (auth server HTTP response code) matches the expecation.

Currently the script is pretty verbose because it is still in
development. At least one test is not working as expected.